+++ minisat/mtl/Alg.h
@@ -35,10 +35,10 @@ namespace Minisat {
 template<class V, class T>
 static inline void remove(V& ts, const T& t)
 {
-    int j = 0;
-    for (; j < (int)ts.size() && ts[j] != t; j++);
-    assert(j < (int)ts.size());
-    for (; j < (int)ts.size()-1; j++) ts[j] = ts[j+1];
+    size_t j = 0;
+    for (; j < ts.size() && ts[j] != t; j++);
+    assert(j < ts.size());
+    for (; j < ts.size()-1; j++) ts[j] = ts[j+1];
     ts.pop();
 }
 
@@ -46,9 +46,9 @@ static inline void remove(V& ts, const T& t)
 template<class V, class T>
 static inline bool find(V& ts, const T& t)
 {
-    int j = 0;
-    for (; j < (int)ts.size() && ts[j] != t; j++);
-    return j < (int)ts.size();
+    size_t j = 0;
+    for (; j < ts.size() && ts[j] != t; j++);
+    return j < ts.size();
 }
 
 
@@ -69,7 +69,7 @@ static inline void copy(const vec<T>& from, vec<T>& to, bool append = false)
 {
     if (!append)
         to.clear();
-    for (int i = 0; i < from.size(); i++){
+    for (size_t i = 0; i < from.size(); i++){
         to.push();
         copy(from[i], to.last());
     }
